drf-tlegram/requirements.txt
pip~=24.1.2
attrs~=23.2.0
pillow~=10.4.0
filelock~=3.15.4
typing_extensions~=4.12.2
idna~=3.7
multidict~=6.0.5
Babel~=2.15.0
pytz~=2024.1
setuptools~=72.1.0
aiohttp~=3.10.0
certifi~=2024.7.4
aiogram~=2.14
yarl~=1.9.4
aiosignal~=1.3.1
frozenlist~=1.4.1
aiohappyeyeballs~=2.3.2
urllib3~=2.2.2
requests~=2.32.3
python-dotenv~=1.0.1